requirements-backend.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
pandas==2.2.3
pdfplumber==0.11.4
PyPDF2==3.0.1
python-multipart==0.0.12
requests==2.32.3
python-dotenv==1.0.1
# Dependências para OCR
PyMuPDF==1.23.26
Pillow==10.1.0
pytesseract==0.3.10